Before dos2unix

+++ tree_destroy.c
@@ -1,9 +1,14 @@
-/*
-* @Author: Work
-* @Date:   2015-11-24 01:04:24
-* @Last Modified by:   Work
-* @Last Modified time: 2015-11-24 02:40:16
-*/
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   tree_destroy.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: nsierra- <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2015/12/04 15:15:05 by nsierra-          #+#    #+#             */
+/*   Updated: 2015/12/04 15:15:07 by nsierra-         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "stdlib.h"
 #include "fttree.h"
@@ -17,8 +22,10 @@ static void			tree_dest_helper(t_tree *tree, t_tree_node *x)
 	{
 		tree_dest_helper(tree, x->left);
 		tree_dest_helper(tree, x->right);
-		tree->destroy_key(x->key);
-		tree->destroy_data(x->data);
+		if (tree->destroy_key)
+			tree->destroy_key(x->key);
+		if (tree->destroy_data)
+			tree->destroy_data(x->data);
 		free(x);
 	}
 }
